requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-encrypted-model-fields==0.5.8
django-formtools==2.2
django-otp==0.9.4
django-phonenumber-field==3.0.1
django-two-factor-auth==1.12.1
djangorestframework==3.11.1
djongo==1.3.3
idna==2.10
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
pyxero==0.9.2
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.10
